+++ src/hooks/useWon.js
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+const useWon = () => {
+  const [won, setWon] = useState(false);
+
+  return { won, setWon };
+};
+
+export default useWon;
